fix(S10): paths of imports
