+++ src/common/types/Product/ProductDescription.ts
