+++ src/common/types/Product/ProductInfo.ts
