+++ src/common/types/Product/index.ts
@@ -1,2 +1,5 @@
 export * from "./Product.ts";
 export * from "./ProductPrice.ts";
+export * from "./ProductInfo.ts";
+export * from "./ProductDescription.ts";
+export * from "./ProductUserFeedback.ts";
