+++ src/common/types/Products.ts
@@ -1,0 +1,10 @@
+import { BaseState } from "./BaseState.ts";
+import { SelectVariants } from "./Select.ts";
+import { Product, ProductCategory } from "./Product";
+
+export type ProductsState = {
+  products: Product[];
+  productsCategoriesWithBrands: ProductCategory[];
+  productsCategories: SelectVariants;
+  searchValue: string;
+} & BaseState;
