--- src/common/types/Products/index.ts
@@ -1,4 +1,0 @@
-export * from "./ProductInfo.ts";
-export * from "./ProductDescription.ts";
-export * from "./ProductUserFeedback.ts";
-export * from "./Products.ts";
