+++ src/common/types/index.ts
@@ -3,8 +3,8 @@ export type { ContextState } from "./ContextState.ts";
 export type { DispatchAction } from "./DispatchAction.ts";
 export type { StoreAsyncThunk } from "./StoreAsyncThunk.ts";
 export type { BaseState } from "./BaseState.ts";
+export type { ProductsState } from "./Products.ts";
 export type * from "./Select.ts";
-export type * from "./Products";
 export type * from "./Product";
 export type * from "./Filters";
 export type * from "./Pagination";
